[ISSUE #11956] Move logback 12 and log4j2 adapter to single module. (#11996)
* Move some basic logging classes into common module.

* Move logback 12 version adapter to single module.

* For checkstyle.

* Enhance logger adapter pom.

* Refactor the adapter initialization logic to avoid spi loader error.

* Move log4j2 logger adapter to single module.

* Fix the ut cover and create useless dir.
