+++ common/src/main/java/com/alibaba/nacos/common/logging/NacosLoggingAdapterBuilder.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 1999-2023 Alibaba Group Holding Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alibaba.nacos.common.logging;
+
+/**
+ * Builder of {@link NacosLoggingAdapter}.
+ *
+ * <p>
+ *     Why not directly SPI {@link NacosLoggingAdapter}?
+ * </p>
+ * <p>
+ *     To avoid some {@link NacosLoggingAdapter} initialization error casue the SPI loader exit.
+ * </p>
+ *
+ * @author xiweng.yy
+ */
+public interface NacosLoggingAdapterBuilder {
+    
+    /**
+     * Build {@link NacosLoggingAdapter} implementation.
+     *
+     * @return {@link NacosLoggingAdapter}
+     */
+    NacosLoggingAdapter build();
+}
