+++ logger-adapter-impl/log4j2-adapter/src/main/java/com/alibaba/nacos/logger/adapter/log4j2/Log4j2NacosLoggingAdapterBuilder.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 1999-2023 Alibaba Group Holding Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alibaba.nacos.logger.adapter.log4j2;
+
+import com.alibaba.nacos.common.logging.NacosLoggingAdapter;
+import com.alibaba.nacos.common.logging.NacosLoggingAdapterBuilder;
+
+/**
+ * Builder of {@link com.alibaba.nacos.common.logging.NacosLoggingAdapter} for log4j2.
+ *
+ * @author xiweng.yy
+ */
+public class Log4j2NacosLoggingAdapterBuilder implements NacosLoggingAdapterBuilder {
+    
+    @Override
+    public NacosLoggingAdapter build() {
+        return new Log4J2NacosLoggingAdapter();
+    }
+}
