+++ logger-adapter-impl/log4j2-adapter/src/main/resources/nacos-log4j2.xml
