+++ logger-adapter-impl/logback-adapter-12/src/test/java/com/alibaba/nacos/logger/adapter/logback12/LogbackNacosLoggingAdapterBuilderTest.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 1999-2023 Alibaba Group Holding Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alibaba.nacos.logger.adapter.logback12;
+
+import com.alibaba.nacos.common.logging.NacosLoggingAdapter;
+import org.junit.Test;
+
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
+
+public class LogbackNacosLoggingAdapterBuilderTest {
+    
+    @Test
+    public void build() {
+        LogbackNacosLoggingAdapterBuilder builder = new LogbackNacosLoggingAdapterBuilder();
+        NacosLoggingAdapter adapter = builder.build();
+        assertNotNull(adapter);
+        assertTrue(adapter instanceof LogbackNacosLoggingAdapter);
+    }
+}
